+++ cpp/command/matchauto.cpp
@@ -441,7 +441,7 @@ int MainCmds::matchauto(int argc, const char* const* argv) {
   //Load per-bot search config, first, which also tells us how many bots we're running
   vector<SearchParams> paramss = Setup::loadParams(cfg,Setup::SETUP_FOR_MATCH);
   assert(paramss.size() > 0);
-  int numBots = paramss.size();
+  int numBots = (int)paramss.size();
 
   //Load the names of the bots and which model each bot is using
   vector<string> nnModelFilesByBot;
